commit on 27-04-2020 at 7:01pm

+++ OnlineTestManagementSystem/src/main/java/com/capgemini/onlinetest/dao/QuestionsRepository.java
@@ -7,6 +7,6 @@ import com.capgemini.onlinetest.entity.Questions;
 
 
 @Repository
-public interface QuestionsRepository extends JpaRepository<Questions,Long>{
+public interface QuestionsRepository extends JpaRepository<Questions,Integer>{
 
 }
