+++ OnlineTestManagementSystem/src/main/java/com/capgemini/onlinetest/dao/TestRepository.java
@@ -8,6 +8,6 @@ import com.capgemini.onlinetest.entity.Test;
 
 
 @Repository
-public interface TestRepository extends JpaRepository<Test,Long>{
+public interface TestRepository extends JpaRepository<Test,Integer>{
 
 }
